Import tkinter for our gui and added the references for the new game menu
